+++ modelmapper/__init__.py
@@ -1,5 +1,5 @@
 # flake8: noqa
-__version__ = '0.4.3'
+__version__ = '0.4.4'
 import sys
 pyversion = float(sys.version[:3])
 if pyversion < 3.6:
@@ -9,5 +9,6 @@ from modelmapper.mapper import Mapper, SqlalchemyFieldType
 from modelmapper.ui import get_user_choice, get_user_input
 from modelmapper.cleaner import Cleaner
 from modelmapper.initialize import initialize
-from modelmapper.loader import Loader, PostgresBulkLoader, PostgresSnapshotLoader
+from modelmapper.etl import ETL
+from modelmapper.loader import PostgresBulkLoaderMixin, PostgresSnapshotLoaderMixin
